New directory structure to move versioning to a branch concern
The current versioning scheme uses a directory per year, and results in copying of data and updating URLs in notebooks. This commit and the new 2025 branch makes versioning a concern of the git branch structure. It is anticipated that each year a new branch will be created named for the year based on the previous year branch. In this way prior year notebooks keep referencing the prior year branches, and we can make changes in place on the new bracnh